advanced_image_processing: Added casting to numpy.uint64 to resolve DeprecationWarning raised by NumPy.

+++ nanshe/advanced_image_processing.py
@@ -1817,7 +1817,7 @@ def postprocess_data(new_dictionary, array_debug_recorder = HDF5_recorder.EmptyA
     if unmerged_neuron_set.size:
         array_debug_recorder("unmerged_neuron_set", unmerged_neuron_set)
 
-        unmerged_neuron_set_contours = unmerged_neuron_set["contour"].copy()
+        unmerged_neuron_set_contours = unmerged_neuron_set["contour"].astype(numpy.uint64).copy()
 
         unmerged_neuron_set_contours *= expanded_numpy.expand_enumerate(unmerged_neuron_set_contours, start = 1)
         unmerged_neuron_set_contours = unmerged_neuron_set_contours.max(axis = 0)
@@ -1827,7 +1827,7 @@ def postprocess_data(new_dictionary, array_debug_recorder = HDF5_recorder.EmptyA
     if new_neurons_set.size:
         array_debug_recorder("new_neurons_set", new_neurons_set)
 
-        new_neurons_set_contours = new_neurons_set["contour"].copy()
+        new_neurons_set_contours = new_neurons_set["contour"].astype(numpy.uint64).copy()
 
         new_neurons_set_contours *= expanded_numpy.expand_enumerate(new_neurons_set_contours, start = 1)
         new_neurons_set_contours = new_neurons_set_contours.max(axis = 0)
